Move download.py to dataset module
